# swagger.yaml
swagger: "2.0"
info:
  description: "este es un ejemplo de como realizar servicios rest con swagger con un .yml y .json"
  version: "1.0.0"
  title: "Swagger Ejemplo de Mi servicio Imundo"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "imundo90@gmail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:8080"
basePath: "/v2"
...
schemes:
- "https"
- "http"
paths:
...
  /docs/findByStatus:
    get:
      tags:
      - "Documentos"
      summary: "encontrar documento by status"
      description: "Multiple status values can be provided with comma separated strings"
      operationId: "findDocumentoByStatus"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "status"
        in: "query"
        description: "Status values that need to be considered for filter"
        required: true
        type: "array"
        items:
          type: "string"
          enum:
          - "available"
          - "pending"
          - "sold"
          default: "available"
        collectionFormat: "multi"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Documentos"
        400:
          description: "Invalid status value"
      security:
      - petstore_auth:
        - "write:documentos"
        - "read:documentos"
  /docs/findByTags:
    get:
      tags:
      - "Documentos"
      summary: "Finds documentos by tags"
      description: "Muliple tags can be provided with comma separated strings. Use tag1, tag2, tag3 for testing."
      operationId: "findDocumentosByTags"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "tags"
        in: "query"
        description: "Tags to filter by"
        required: true
        type: "array"
        items:
          type: "string"
        collectionFormat: "multi"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
             $ref: "#/definitions/Documentos"
        400:
          description: "Invalid tag value"
      security:
      - petstore_auth:
        - "write:documentos"
        - "read:documentos"
      deprecated: true
  /docs/{docId}:
    get:
      tags:
      - "Documentos"
      summary: "Find doc by ID"
      description: "Returns a single doc"
      operationId: "getDocumentosById"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "docId"
        in: "path"
        description: "ID of pet to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Documentos"
        400:
          description: "Invalid ID supplied"
        404:
          description: "Doc not found"
      security:
      - api_key: []
    post:
      tags:
      - "Documentos"
      summary: "Updates a pet in the store with form data"
      description: ""
      operationId: "updateDocWithForm"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "docId"
        in: "path"
        description: "ID of doc that needs to be updated"
        required: true
        type: "integer"
        format: "int64"
      - name: "name"
        in: "formData"
        description: "Updated name of the doc"
        required: false
        type: "string"
      - name: "status"
        in: "formData"
        description: "Updated status of the doc"
        required: false
        type: "string"
      responses:
        405:
          description: "Invalid input"
      security:
      - petstore_auth:
        - "write:documentos"
        - "read:documentos"
    delete:
      tags:
      - "Documentos"
      summary: "Deletes a documentos"
      description: ""
      operationId: "deleteDocumentos"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "api_key"
        in: "header"
        required: false
        type: "string"
      - name: "docId"
        in: "path"
        description: "Doc id to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        400:
          description: "Invalid ID supplied"
        404:
          description: "Doc not found"
      security:
      - petstore_auth:
        - "write:documentos"
        - "read:documentos"
  /docs/{docId}/uploadImage:
    post:
      tags:
      - "Documentos"
      summary: "uploads an image"
      description: ""
      operationId: "uploadFile"
      consumes:
      - "multipart/form-data"
      produces:
      - "application/json"
      parameters:
      - name: "docId"
        in: "path"
        description: "ID of doc to update"
        required: true
        type: "integer"
        format: "int64"
      - name: "additionalMetadata"
        in: "formData"
        description: "Additional data to pass to server"
        required: false
        type: "string"
      - name: "file"
        in: "formData"
        description: "file to upload"
        required: false
        type: "file"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
      security:
      - petstore_auth:
        - "write:documentos"
        - "read:documentos"
...
securityDefinitions:
  petstore_auth:
    type: "oauth2"
    authorizationUrl: "http://petstore.swagger.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write:documentos: "modify documentos in your account"
      read:documentos: "read your documentos"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
...
  Category:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
    xml:
      name: "Category"
...
  Tag:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
    xml:
      name: "Tag"
  Documentos:
    type: "object"
    required:
    - "name"
    - "photoUrls"
    properties:
      id:
        type: "integer"
        format: "int64"
      category:
        $ref: "#/definitions/Category"
      name:
        type: "string"
        example: "doggie"
      photoUrls:
        type: "array"
        xml:
          name: "photoUrl"
          wrapped: true
        items:
          type: "string"
      tags:
        type: "array"
        xml:
          name: "tag"
          wrapped: true
        items:
          $ref: "#/definitions/Tag"
      status:
        type: "string"
        description: "pet status in the store"
        enum:
        - "available"
        - "pending"
        - "sold"
    xml:
      name: "Documentos"
  ApiResponse:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      type:
        type: "string"
      message:
        type: "string"
...